+++ supplier/serializers.py
@@ -54,7 +54,8 @@ class SupplierSerializer(serializers.ModelSerializer):
             'date_joined',
             'sso_id',
             'is_company_owner',
-            'role'
+            'role',
+            'name',
         )
         extra_kwargs = {
             'sso_id': {'required': True},