+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Pre release
 
+### Implemented enhancements
+- TT-1733 - Added request identity verification feature
+- TT-1733 - Allow name to be provided on supplier create
+
 ### Fixed bugs
 - No ticket - Upgrade django to 1.11.23 to fix vulnerability